Fix campaign tabs on place listing page
